acho que finalizei

+++ src/main/java/EditorDeVideo.java
@@ -1,11 +1,12 @@
-public class EditorDeVideo extends Funcionario {
-
-    public EditorDeVideo(String nome, String numeroCpf, String cargo, float salario) {
+public class EditorDeVideo extends Funcionario
+{
+    public EditorDeVideo(String nome, String numeroCpf, String cargo, float salario)
+    {
         super(nome, numeroCpf, cargo, salario);
     }
 
     public float calculaBonificacao()
     {
-        return 600;
+        return (float) 600;
     }
 }
